Исправил часть ошибок структуры, на которые было указано.

+++ Assets/Scripts/Data/UpgradeItemConfig.cs
@@ -10,13 +10,13 @@ public class UpgradeItemConfig : ScriptableObject
     private UpgradeType _upgradeType;
 
     [SerializeField]
-    private int _valueUpgrade;
+    private float _valueUpgrade;
 
     public int Id => _itemConfig.Id;
 
     public UpgradeType UpgradeType => _upgradeType;
 
-    public int ValueUpgrade => _valueUpgrade;
+    public float ValueUpgrade => _valueUpgrade;
 }
 
 public enum UpgradeType
